Added 100+ series
Methods rubric has been updated. 

All methods that end with 1,2,3 had user behaviours as previously discussed. This has been changed to only be the case for methods <100. Since we are expanding the range of values possible for method, it has also been changed from a u8 -> u32. 

Now, user behaviours have NOT been excluded from the 100+ series, they are instead indicated by the whole number division with 100. So for eg, methods 300-399, all encompass user behaviour 3 (method %3 below 100) as before. Thus, anything like 400, or 500 will work but will simply not have any user behaviour that prompts the code to "break" out of the loop that is rolling until success.

100+ series replaced the %10 logic to indicate method user behaviour and instead indicate, a different type of system for the trials - positive reinforcement system. This is a system that rewards the player for being lucky. So the %10 digit for 100+ methods now indicates the number of successive successes after which, a free trial success is provided. For instance, method 112 simulates a user using method 1, but since 112%10 = 2, it also gives a free trial success after 2 successive passes. Understandably this provides users the option to set it to something like say 119, so you would only get a free pass after 9 successive successes. Note that by design, methods 119 and 129, should not have any difference. 

One more important note however, is that we have provided users the option to also simulate this positive reinforcement method TOGETHER with pity system where you get a free pass after 2 fails. That has been set to be available for the methods 100-109. This is done by replacing the initial line method < 10 to method %100 < 10. Hence, by design, this would mean that something like method 209, 302, 405 for eg, would all use a positive reinforcement system, AND the pity system as outlined.

Have decided against adding the same level of control to pity system (eg, success after 3 consecutive fails instead of 2, purely because it does not seem logical to want to simulate this ... unless desired). Also, we have decided against making it possible for the user to instead set something like 122 being a simulation whereby the user could indicate that the trial system give a free pass after 22 successive successes. This decision should be understandable.