Pity system inclusion in 100+ series
Has been fully integrated. Previously, pity system could be triggered in 100 series whereby %100<10. But the double upping of pity system at fixed levels INCLUDING the first level (so this affected trials even without any manually fixed (*) stages) was not accounted for. In the typical pity system incorporations, that is REMOVED. 

In our trial simulator, it is meant to exclude it too for methods 1-4 and now, it is discounted for any methods above 100 where %100<5 equivalently. This modularity allows for robust comparison between various situations.